Add pagination and  Write unitests for sort feature

+++ app/services/collaboration_services/remove_collaborator.rb
@@ -3,20 +3,23 @@ module CollaborationServices
     include BaseService
 
     def initialize(params)
-      @challenge = Challenge.find(params[:id])
+      @challenge_id = params[:id]
       @user = params[:user]
     end
 
     def call
+      raise StandardError if @user.nil? || @challenge_id.nil?
+
+      @challenge = Challenge.find(@challenge_id)
       Collaboration.where(challenge: @challenge, user: @user).destroy_all
     rescue StandardError => e
       Failure.new({ message: e.message })
     else
-      Success.new({})
+      Success.new({ message: 'Your collaboration has been removed for this challenge.' })
     end
 
     private
 
-    attr_reader :challenge, :user
+    attr_reader :challenge, :challenge_id, :user
   end
 end
